Bug fix with time comparison
I wrongly assumed I have to add the offset to keep the observations to the current time. Instead it should be applied to the firstRecorded time that is stored in DynamoDB. When the current time exceeds this time plus the offset, then delete the items that have the station id as the key.